src/backend/list_models.py
import os
from dotenv import load_dotenv
import google.generativeai as genai

load_dotenv(encoding="utf-8")
genai.configure(api_key=os.environ.get("GEMINI_API_KEY"))

print("Listing available models...")
try:
    for m in genai.list_models():
        if 'generateContent' in m.supported_generation_methods:
            print(m.name)
except Exception as e:
    print(f"Error: {e}")
