src/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
numpy
python-dotenv
requests
slowapi
pydantic
groq
deepgram-sdk

# Supabase & Authentication
supabase
PyJWT
bcrypt

# Resume Parsing
PyPDF2
python-docx
pdfplumber

# ATS Analysis
google-generativeai
scikit-learn
